fix: circular dependency for REGISTRY_TYPE_MAP fix
