+++ src/Shared/Components/RegistryIcon/index.ts
@@ -1,1 +1,2 @@
 export * from './RegistryIcon'
+export * from './constants'
